backend/requirements.txt
openai>=0.28,<1.0.0
telethon>=1.26.0
python-dotenv>=1.0.0
asyncio>=3.4.3
